Move RequestDelegateFactory tests to be shared with RDG
